Fixing FileSymbolicLink to work with Scons newer than 2.3.6
In Scons 2.3.6 they have added __slots__ to Node objects in
order to reduce memory footprint.
Our FileSymbolicLink implementation must also support the same slots
Since we add _linkto additional attribute to our node and there is
no slot for it we add __linktos__ dictionary to our class to store
the attribute for each instance of FileSymbolicLink.